delete mushroom is working
