+++ client/src/App.js
@@ -5,9 +5,12 @@ import {
   Route,
   Redirect,
 } from "react-router-dom";
+import API from "./utils/API";
 
 // Import Contexts
 import AuthenticationContext from "./contexts/AuthenticationContext";
+import MushroomContext from "./contexts/MushroomContext";
+
 
 // Import Components
 import Navbar from "./components/Nav/Navbar";
@@ -16,27 +19,32 @@ import Navbar from "./components/Nav/Navbar";
 import Home from "./pages/home";
 import About from "./pages/about";
 import Shop from "./pages/Shop";
-import Signup from "./pages/signup";
-import Login from "./pages/login";
-import AdminLogin from "./pages/AdminLogin";
-import AdminDashboard from "./pages/AdminDashboard";
-import AdminProducts from "./pages/AdminProducts";
+import Signup from "./pages/Authentication/Signup";
+import Login from "./pages/Authentication/Login";
+import AdminLogin from "./pages/Admin/Login";
+import AdminDashboard from "./pages/Admin/Dashboard";
+import AdminProducts from "./pages/Admin/Products";
 import Members from "./pages/Members";
 import NoMatch from "./pages/NoMatch";
 
 const App = () => {
+  // Authentication State Variables
   const [isAdmin, setIsAdmin] = useState(false);
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [token, setToken] = useState(null);
-  const [userID, setUserID] = useState(null);   
+  const [userID, setUserID] = useState(null);
+  
+  // Mushroom State Variables
+  const [mushrooms, setMushrooms] = useState([]);
 
-  // useEffect(() => {
-  //   setIsAdmin(false);
-  //   setIsAuthenticated(false);
-  //   setToken(null);
-  //   setUserID(null);
-  // }, []);
+  // Mushroom Functions for use with MushroomContext
+  function getMushrooms(){
+    API.getMushrooms().then((res) => {
+      setMushrooms(res.data);
+    });
+  }
 
+  // Authentication Functions for use with AuthenticationContext
   const login = useCallback((userID, isAdmin, token) => {
     setIsAdmin(isAdmin);
     setUserID(userID);
@@ -124,10 +132,12 @@ const App = () => {
         logout: logout
       }}
     >
+    <MushroomContext.Provider value={{mushrooms: mushrooms, getMushrooms: getMushrooms}}>
       <Router>
         <Navbar />
         {routes}
       </Router>
+      </MushroomContext.Provider>
     </AuthenticationContext.Provider>
   );
 };
