+++ client/src/pages/Authentication/Login/login.css
@@ -1,0 +1,28 @@
+.general-form{
+  width: 450px;
+  margin: auto;
+  background: #ffffff;
+  box-shadow: 0px 14px 80px rgba(34, 35, 58, 0.2);
+  padding: 40px 55px 45px 55px;
+  border-radius: 15px;
+  transition: all .3s;
+}
+
+.forgot-password,
+.forgot-password a {
+  float: right;
+  font-size: 13px;
+  padding-top: 10px;
+  color: #7f7d7d;
+  margin: 0;
+}
+
+.forgot-password a {
+  color: #167bff;
+  padding-top:0;
+  padding-left: 5px;
+}
+
+.text-left{
+  float:left;
+}
