+++ client/src/utils/API.js
@@ -11,7 +11,7 @@ export default {
   },
   // Deletes the mushroom with the given id
   deleteMushroom: function (id) {
-    return axios.delete("/api/mushroom/" + id);
+    return axios.delete("/api/mushrooms/" + id);
   },
   // Saves a mushroom to the database
   saveMushroom: function (mushroomData) {
